requirements.txt
boto3==1.41.0
Django==5.2.8
django-storages==1.14.6
gunicorn==23.0.0
psycopg2-binary==2.9.11
python-decouple==3.8
